move key change function to new file
