+++ dexios/src/main.rs
@@ -52,7 +52,7 @@ fn main() -> Result<()> {
 
                 let (keyfile_old, keyfile_new) = key_update_params(sub_matches_update_key)?;
 
-                subcommands::header::update_key(
+                subcommands::header_key::update_key(
                     &get_param("input", sub_matches_update_key)?,
                     &keyfile_old,
                     &keyfile_new,
